Revert "rpc: adjust `getversion` RPC response"
This reverts commit 1c6afe402fb60c371490fd096f1d8abfff666d63.

+++ pkg/rpc/server/server.go
@@ -551,7 +551,7 @@ func (s *Server) getVersion(_ request.Params) (interface{}, *response.Error) {
 			MaxTransactionsPerBlock:     cfg.MaxTransactionsPerBlock,
 			MemoryPoolMaxTransactions:   cfg.MemPoolSize,
 			ValidatorsCount:             byte(cfg.GetNumOfCNs(s.chain.BlockHeight())),
-			InitialGasDistribution:      int64(cfg.InitialGASSupply),
+			InitialGasDistribution:      cfg.InitialGASSupply,
 			StateRootInHeader:           cfg.StateRootInHeader,
 		},
 	}, nil
